refactor: merge small utils into utils files
