--- packages/core-database-postgres/src/utils/camelize-columns.ts
@@ -1,17 +1,0 @@
-/* tslint:disable:forin prefer-for-of*/
-
-export function camelizeColumns(pgp, data) {
-    const tmp = data[0];
-
-    for (const prop in tmp) {
-        const camel = pgp.utils.camelize(prop);
-
-        if (!(camel in tmp)) {
-            for (let i = 0; i < data.length; i++) {
-                const d = data[i];
-                d[camel] = d[prop];
-                delete d[prop];
-            }
-        }
-    }
-}
