--- packages/core-database-postgres/src/utils/index.ts
@@ -1,4 +1,0 @@
-import { camelizeColumns } from "./camelize-columns";
-import { loadQueryFile } from "./load-query-file";
-
-export { camelizeColumns, loadQueryFile };
