--- packages/crypto/src/utils/is-exception.ts
@@ -1,12 +1,0 @@
-import { configManager } from "../managers";
-import { IBlockData, ITransactionData } from "../models";
-
-/**
- * Check if the given block or transaction id is an exception.
- */
-export function isException(blockOrTransaction: IBlockData | ITransactionData): boolean {
-    return ["blocks", "transactions"].some(key => {
-        const exceptions = configManager.get(`exceptions.${key}`);
-        return Array.isArray(exceptions) && exceptions.includes(blockOrTransaction.id);
-    });
-}
